refactor: move fetch helpers
