+++ packages/common/src/config.ts
@@ -2,6 +2,7 @@
 
 /**
  * @ignore
+ * todo: remove this file (needed for CLI?)
  */
 const config = {
   // network: network.defaults.MAINNET_DEFAULT,
