+++ packages/common/src/index.ts
@@ -7,4 +7,4 @@ export * from './signatures';
 export * from './keys';
 export * from './buffer';
 export * from './types';
-export * from './contract-abi';
+export * from './fetch';
