+++ packages/transactions/src/constants.ts
@@ -1,14 +1,31 @@
+/** @ignore internal */
+export const BLOCKSTACK_DEFAULT_GAIA_HUB_URL = 'https://hub.blockstack.org';
+
 /**
  * The chain ID (unsigned 32-bit integer), used so transactions can't be replayed on other chains.
  * Similar to the {@link TransactionVersion}.
  */
-export enum ChainID {
+export enum ChainId {
   Testnet = 0x80000000,
   Mainnet = 0x00000001,
 }
-// todo: deduplicate chain id
 
-export const DEFAULT_CHAIN_ID = ChainID.Mainnet;
+/**
+ * The **peer** network ID.
+ * Typically not used in signing, but used for broadcasting to the P2P network.
+ * It can also be used to determine the parent of a subnet.
+ *
+ * **Attention:**
+ * For mainnet/testnet the v2/info response `.network_id` refers to the chain ID.
+ * For subnets the v2/info response `.network_id` refers to the peer network ID and the chain ID (they are the same for subnets).
+ * The `.parent_network_id` refers to the actual peer network ID (of the parent) in both cases.
+ */
+export enum PeerNetworkId {
+  Mainnet = 0x17000000,
+  Testnet = 0xff000000,
+}
+
+export const DEFAULT_CHAIN_ID = ChainId.Mainnet;
 export const MAX_STRING_LENGTH_BYTES = 128;
 export const CLARITY_INT_SIZE = 128;
 export const CLARITY_INT_BYTE_SIZE = 16;
@@ -108,7 +125,7 @@ export function anchorModeFrom(mode: AnchorModeName | AnchorMode): AnchorMode {
 
 /**
  * The transaction version, used so transactions can't be replayed on other networks.
- * Similar to the {@link ChainID}.
+ * Similar to the {@link ChainId}.
  * Used internally for serializing and deserializing transactions.
  */
 export enum TransactionVersion {
@@ -117,6 +134,11 @@ export enum TransactionVersion {
 }
 
 export const DEFAULT_TRANSACTION_VERSION = TransactionVersion.Mainnet;
+
+/** @ignore */
+export function whenTransactionVersion(transactionVersion: TransactionVersion) {
+  return <T>(map: Record<TransactionVersion, T>): T => map[transactionVersion];
+}
 
 /**
  * How to treat unspecified transfers of a transaction.
@@ -219,7 +241,7 @@ export enum NonFungibleConditionCode {
 /**
  * The type of sender for a post-condition.
  */
-export enum PostConditionPrincipalID {
+export enum PostConditionPrincipalId {
   Origin = 0x01,
   Standard = 0x02,
   Contract = 0x03,
