+++ packages/transactions/tests/fetch.test.ts
@@ -1,5 +1,5 @@
+import { fetchWrapper, getFetchOptions, setFetchOptions } from '@stacks/common';
 import fetchMock from 'jest-fetch-mock';
-import { fetchWrapper, getFetchOptions, setFetchOptions } from '../src';
 
 test('Verify fetch private options', async () => {
   const defaultOptioins = getFetchOptions();
