+++ packages/wallet-sdk/src/utils.ts
@@ -1,7 +1,7 @@
-import { bytesToHex, ChainID } from '@stacks/common';
+import { bytesToHex } from '@stacks/common';
 import { getPublicKeyFromPrivate, publicKeyToBtcAddress, randomBytes } from '@stacks/encryption';
-import { createFetchFn, FetchFn } from '@stacks/network';
 import { GaiaHubConfig } from '@stacks/storage';
+import { ChainId, FetchFn, createFetchFn } from '@stacks/transactions';
 import { Json, TokenSigner } from 'jsontokens';
 import { parseZoneFile } from 'zone-file';
 
@@ -125,9 +125,9 @@ export const makeGaiaAssociationToken = ({
 };
 
 interface WhenChainIdMap<T> {
-  [ChainID.Mainnet]: T;
-  [ChainID.Testnet]: T;
+  [ChainId.Mainnet]: T;
+  [ChainId.Testnet]: T;
 }
-export function whenChainId(chainId: ChainID) {
+export function whenChainId(chainId: ChainId) {
   return <T>(chainIdMap: WhenChainIdMap<T>): T => chainIdMap[chainId];
 }
